2007-03-23  Jim Evins  <evins@snaught.com>
	Merged all changes from experimental cairo branch to trunk.
	Trunk no longer depends on libgnomecanvas or libgnomeprint[ui].


git-svn-id: https://glabels.svn.sourceforge.net/svnroot/glabels/trunk@639 f5e0f49d-192f-0410-a22d-a8d8700d0965

+++ glabels2/src/merge.c
@@ -123,7 +123,7 @@ gl_merge_register_backend (GType              type,
 		class = g_type_class_ref (type);
 		if (class == NULL) {
 			g_message ("%s: unknown object type %d",
-				   G_STRLOC, type);
+				   G_STRLOC, (int)type);
 			break;
 		}
 		pspec = g_object_class_find_property (class, pname);
